Water Metering, some preparations.

+++ src/main/java/org/thingsboard/trendz/generator/solution/watermetering/model/City.java
@@ -17,7 +17,7 @@ public class City implements ModelEntity {
     private String systemName;
     private String systemLabel;
 
-    private int population;
+    private long population;
 
     private Set<Region> regions;
     private Set<PumpStation> pumpStations;
